Moved ConcurrentLruSet<T> into Sparrow.
