+++ Raven.Client.Lightweight/Raven.Client.Lightweight.csproj
@@ -213,7 +213,6 @@
     <Compile Include="Connection\DocumentConventionJsonExtensions.cs" />
     <Compile Include="Connection\JsonLuceneDateTimeConverter.cs" />
     <Compile Include="Connection\MultiGetOperation.cs" />
-    <Compile Include="Connection\Profiling\ConcurrentLruLSet.cs" />
     <Compile Include="Connection\Profiling\IHoldProfilingInformation.cs" />
     <Compile Include="Connection\Profiling\ProfilingContext.cs" />
     <Compile Include="Connection\Profiling\RequestStatus.cs" />
