+++ Raven.Sparrow/Sparrow/Hashing.cs
@@ -5,7 +5,7 @@ using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Voron.Util
+namespace Sparrow
 {
     public unsafe static class Hashing
     {
