+++ Raven.Sparrow/Sparrow/readme.md
@@ -1,0 +1,5 @@
+﻿# Sparrow
+
+Sparrow is a collection of utility classes focused on high-performance currently being developed and used in RavenDB and Voron.
+
+The criteria for inclusion of an utility class in Sparrow is that it is dependency free from any external package.
