+++ Raven.Voron/Voron.Tests/Voron.Tests.csproj
@@ -159,7 +159,6 @@
     <Compile Include="Trees\TreeRenaming.cs" />
     <Compile Include="Trees\Updates.cs" />
     <Compile Include="Util\CrcTests.cs" />
-    <Compile Include="Util\Hashing.cs" />
     <Compile Include="Util\ImmutableAppendOnlyTests.cs" />
     <Compile Include="Util\StreamExtensions.cs" />
   </ItemGroup>
