+++ packages/repositories.config
@@ -14,6 +14,7 @@
   <repository path="..\Raven.Server\packages.config" />
   <repository path="..\Raven.SlowTests\packages.config" />
   <repository path="..\Raven.Smuggler\packages.config" />
+  <repository path="..\Raven.Sparrow\Sparrow.Tests\packages.config" />
   <repository path="..\Raven.Studio.Html5\packages.config" />
   <repository path="..\Raven.Tests.Bundles\packages.config" />
   <repository path="..\Raven.Tests.Common\packages.config" />
